requirements.txt
# AI and Machine Learning
google-generativeai>=0.3.0

# Google Services
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Document Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.0.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Data Validation and Models
pydantic>=2.0.0
pydantic[email]>=2.0.0
email-validator>=2.0.0

# Dashboard
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.1.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.12.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
pyyaml>=6.0.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Logging and Monitoring
python-json-logger>=2.0.7